+++ blocks/Illuminator.h
@@ -22,7 +22,7 @@
 
 #include "blocks/Active.h"
 
-const ushort MAX_FUEL = SECONDS_IN_DAY;
+const int MAX_FUEL = SECONDS_IN_DAY;
 
 class Illuminator : public Active {
     Q_OBJECT
@@ -33,7 +33,7 @@ public:
     int     ShouldAct() const override;
     int     LightRadius() const override;
     bool    ShouldFall() const override;
-    void    Damage(ushort dmg, int dmg_kind) override;
+    void    Damage(int dmg, int dmg_kind) override;
     quint8  Kind() const override;
     Block * DropAfterDamage() override;
     QString FullName() const override;
